Add support for message replies
Messages with a reply now have a small indicator to show to which user
the message was in reply to, with the message text in the hover card
for that user.

If the message content is irretrievable (text and sending user), then
it just notes that the user is unknown.