Add workaround for FML channel list issue
On Forge-Forge connections, the vanilla channel list contains the
existence channel, even if the remote side doesn't register the channel.

To fix this, we check if the FML connection data is present and, if so,
use that as our sole source of info for whether the channel is present
or not. We then fall back to #isRemotePresent (for the vanilla channel
list).